fix(protocol): five defects found by the new protocol e2e suites
All five were caught by true end-to-end suites (client vs independently
implemented in-test servers on real loopback sockets) that the previous
unit-only gates could not see.

- dialer/naive: randInt reduced in the uint32 domain. Converting the raw
  draw to int first wraps negative on 32-bit platforms for half of the
  draws, and Go's % keeps the sign, so the padding codec panicked
  (paddingChars[-34]) and produced negative sizes on arm/386/mips.
- protocol/vmess: unmap v4-in-v6 UDP AddrPorts at both storage sites.
  net.ResolveUDPAddr returns a 16-byte mapped IP, so ReadFrom reported
  [::ffff:127.0.0.1]:p and the packet-addr path encoded an IPv4 target
  with the IPv6 addr type.
- protocol/shadowsocks_2022: verify the request-salt echo in the response
  fixed-length header, as SIP022 3.1.3 requires. The parser skipped the
  field, so a response stream was not bound to its request; mismatch now
  fails with protocol.ErrFailAuth.
- protocol/http: return the connect candidate as the netproxy.Conn
  interface, not *http2Conn. A typed nil *http2Conn wrapped into a
  non-nil interface made discardCandidate Close() a nil receiver and
  segfaulted the dial path when the proxy denied CONNECT over h2.
- transport/simpleobfs: search readFirstResponse's accumulated buffer
  for the terminator before applying the size bound. One read carrying
  the full header plus body bytes (the normal fast-relay case) aborted
  with "response header exceeds 8192 bytes" about half the time.

+++ dialer/naive/padding.go
@@ -73,7 +73,11 @@ func randInt(max int) int {
 		// fast PRNG instead of deterministic zero padding.
 		return fastrand.Intn(max)
 	}
-	return int(binary.BigEndian.Uint32(buf[:])) % max
+	// Reduce in the uint32 domain. Converting the raw draw to int first
+	// wraps negative on 32-bit platforms for half of the draws (uint32 with
+	// the top bit set), and Go's % keeps the sign of the dividend, so the
+	// result fed negative sizes and indices into the padding codec.
+	return int(binary.BigEndian.Uint32(buf[:]) % uint32(max))
 }
 
 // paddedWriter wraps an io.Writer and applies naiveproxy payload padding
